requirements.txt
edge-tts>=7.0.0
requests>=2.31.0
Pillow>=10.0.0
